Mise en forme des résultats
Intégration du graphe dans la notice détaillée
Filtres Responsabilités

+++ UnimarcLRM/generate_results.py
@@ -2,9 +2,13 @@
 
 import os
 from unidecode import unidecode
+from lxml.html import parse
+from lxml import etree
 from common_dicts import *
 
 from random import choice
+
+from generate_graph import oeuvreid2graph
 
 def clean_str(string):
     string = unidecode(string.lower())
@@ -68,7 +72,6 @@ def write_html_full_body(file, recordid, record, query, i, dict_entities):
     file.write("\n<body>\n")
     file.write(generate_entete(query, i))
     file.write(f"\n<p>{record.detailed}</p>")
-    file.write(f"\n<p class='recordid'>{recordid}</p>")
 
     # ----------------------- #
     #    Les filtres          #
@@ -80,7 +83,14 @@ def write_html_full_body(file, recordid, record, query, i, dict_entities):
     #    Les exemplaires      #
     # ------------------------#
     div_items = generate_html_items(record, dict_entities)
-    file.write(f"\n<div class='iteùs'>{div_items}</div>")
+    file.write(f"\n<div class='items'>{div_items}</div>\n")
+
+
+    # ------------------------#
+    #   Les infos pro         #
+    # ------------------------#
+    div_infos_pro = generate_pro_infos(recordid, record, dict_entities)
+    file.write(f"\n<div class='pro'>{div_infos_pro}</div>\n")
 
     file.write("\n</body>\n")
 
@@ -107,9 +117,29 @@ def generate_html_items(record, dict_entities):
     return div_items
 
 
+def generate_pro_infos(recordid, record, dict_entities):
+    # Renvoi d'un ensemble d'infos professionnelles
+    infos_pro = "<hr/>\n<h3>Infos professionnelles</h3>"
+    infos_pro += f"\n<div class='recordid'><p class='recordid'>{recordid}</p><//div>\n"
+
+    graph = oeuvreid2graph(record, dict_entities, size="20,20")
+    svg_graph = f"graphs/{recordid}.gv.svg"
+    # div_graph = f'<a target="blank" href="{svg_graph}" title="Afficher le graphe"><img src = "{svg_graph}" alt="Graphe de l\'oeuvre"/>'
+    div_graph = f'\n<div class="graph">{graphfilename2content(svg_graph)}</div>\n'
+    infos_pro += div_graph
+    return f'\n{infos_pro}\n'
+    
+def graphfilename2content(graphfilename):
+    svg_content = etree.parse(os.path.join("results", graphfilename))
+    svg_content = svg_content.xpath("//s:svg", namespaces={"s": "http://www.w3.org/2000/svg"})[0]
+    svg_content = etree.tostring(svg_content).decode("utf-8")
+    return svg_content
+
 def generate_work_filters(record):
     # A partir d'une notice d'oeuvre, générer une liste de liens
     # (<a href=''>) permettant de gérer des filtres
+    
+    # Première ligne de filtres (moteur de recherche, langues, dates)
     filter_form = "<input type='form' width='150px' value='Limiter les résultats'/>"
     filters_lng = ""
     for lang in record.lang:
@@ -119,7 +149,16 @@ def generate_work_filters(record):
             link = f'<a class="filters_lng" href="#">{lang}</a>'
         filters_lng += f" {link}"
     filters_elements = [filter_form, filters_lng]
-    return " ".join(filters_elements)
+    filters1 = f"<div class='filters1'>{' '.join(filters_elements)}</div>"
+
+    # Deuxième ligne (mentions de responsabilités des expressions)
+    links_resp = []
+    for resp in record.exprResp:
+        link = f'<a href="#">{resp}</a>'
+        links_resp.append(link)
+    links_resp = " ".join(links_resp)
+    filters2 = f"<div class='filters2'>{links_resp}</div>"
+    return "\n".join([filters1, filters2])
 
 
 def generate_entete(query, no_resultat=0):
@@ -130,7 +169,7 @@ def generate_entete(query, no_resultat=0):
         entete += f'<h1>Résultat {str(no_resultat)}</h1>'
     else:
         entete += "<h1>Liste des résultats</h1>"
-    entete += "<div>"
+    entete += "</div>"
     return entete
 
 
@@ -150,4 +189,7 @@ def write_html_footer(file):
         
 def delete_html_results():
     for filename in os.listdir("results"):
-        os.remove(f"results/{filename}")
+        if ".html" in filename:
+            os.remove(f"results/{filename}")
+    for filename  in os.listdir("results/graphs"):
+            os.remove(f"results/graphs/{filename}")
